Moved getTLA() to autocad.js for general usage
